pimod gain (WIP)

+++ mt-kahypar/datastructures/hypergraph_common.h
@@ -57,7 +57,7 @@ using HypernodeID = uint32_t;
 using HyperedgeID = uint32_t;
 #endif
 using HypernodeWeight = int32_t;
-using HyperedgeWeight = int32_t;
+using HyperedgeWeight = int64_t;
 using PartitionID = int32_t;
 using Gain = HyperedgeWeight;
